requirements.txt
uvicorn==0.27.1
fastapi
requests==2.31.0
openai
python-dotenv
streamlit
langfuse